Simona2mdf should work without having to specify a directory name
svn path=/trunk/matlab/; revision=10917


Former-commit-id: fd38dfae6f622f1e1e3d77f196827b9b3b93c308

+++ applications/delft3d/nest_matlab/simona2mdf/simona2mdf_area.m
@@ -26,7 +26,9 @@ end
 %
 
 if ~isempty(mdf.filcco)
-   copyfile([mdf.pathsimona filesep mdf.filcco],[mdf.pathd3d filesep simona2mdf_rmpath(mdf.filcco)]);
+    if ~exist([mdf.pathd3d filesep simona2mdf_rmpath(mdf.filcco)],'file');
+        copyfile([mdf.pathsimona filesep mdf.filcco],[mdf.pathd3d filesep simona2mdf_rmpath(mdf.filcco)]);
+    end  
 %  mdf        = rmfield(mdf,'pathsimona');
 %  mdf        = rmfield(mdf,'pathd3d');
    mdf.filcco = simona2mdf_rmpath(mdf.filcco);
